dom/indexedDB/test/test_open_empty_db.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript">
    function run_test()
    {
      const name = window.location.pathname;
      const description = "My Test Database";

      function errorHandler(event) {
        ok(false, "indexedDB error (" + event.code + "): " + event.message);
        SimpleTest.finish();
      }

      var request;

      try {
        request = indexedDB.open("", "");
        ok(false, "Open with empty name should have thrown!");
      }
      catch(e) {
        is(request, undefined, "Shouldn't be set to anything");
      }

      try {
        request = indexedDB.open(null, "");
        ok(false, "Open with null name should have thrown!");
      }
      catch(e) {
        is(request, undefined, "Shouldn't be set to anything");
      }

      request = indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = function(event) {
        ok(event.source == indexedDB, "event.source is correct");
        var db = event.result;
        is(db.name, name, "Bad name");
        is(db.description, description, "Bad description");
        is(db.version, "", "Bad version");

        var count = db.objectStores.length;
        is(count, 0, "Bad objectStores list");

        count = db.indexes.length;
        is(count, 0, "Bad indexes list");

        is(db.currentTransaction, null, "Bad transaction");

        SimpleTest.finish();
      };
    }

    SimpleTest.waitForExplicitFinish();
  </script>

</head>

<body onload="run_test();">
</body>
</html>
